requirements.txt
# Build tools (required for building packages from source)
setuptools>=68.0.0
wheel>=0.40.0

# Core dependencies
python-telegram-bot==20.7
supabase==2.3.4
python-dotenv==1.0.0

# OpenAI integration
openai==1.12.0

# LangChain and LangGraph

# Audio processing
soundfile==0.12.1
numpy>=1.26.0

# HTTP requests
requests==2.31.0

# Caching
cachetools==5.3.2

# Web server for health checks
aiohttp==3.9.1

# Type hints (if using older Python versions)
typing-extensions==4.8.0
